+++ fdr-stack/src/main/java/com/dwarfeng/fdr/stack/handler/MapLocalCacheHandler.java
@@ -1,0 +1,33 @@
+package com.dwarfeng.fdr.stack.handler;
+
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
+import com.dwarfeng.subgrade.stack.handler.Handler;
+
+/**
+ * 映射本地缓存处理器。
+ *
+ * <p>处理器在本地保存数据，缓存中的数据可以保证与数据源保持同步。</p>
+ * <p>数据存放在本地，必要时才与数据访问层通信，这有助于程序效率的提升。</p>
+ * <p>该处理器线程安全。</p>
+ *
+ * @author DwArFeng
+ * @since 1.9.0
+ */
+public interface MapLocalCacheHandler extends Handler {
+
+    /**
+     * 获取指定类型的映射器。
+     *
+     * @param mapperType 指定的映射器类型。
+     * @return 指定的映射器，或者是null。
+     * @throws HandlerException 处理器异常。
+     */
+    Mapper getMapper(String mapperType) throws HandlerException;
+
+    /**
+     * 清除本地缓存。
+     *
+     * @throws HandlerException 处理器异常。
+     */
+    void clear() throws HandlerException;
+}