+++ fdr-stack/src/main/java/com/dwarfeng/fdr/stack/service/MappingLookupService.java
@@ -17,50 +17,47 @@ import java.util.List;
 public interface MappingLookupService extends Service {
 
     /**
-     * 查询指定数据点的实时数据。
-     *
-     * @param pointKey 指定的数据点。
-     * @return 指定数据点的实时数据。
-     * @throws ServiceException 服务异常。
-     */
-    TimedValue lookupRealtime(LongIdKey pointKey) throws ServiceException;
-
-    /**
-     * 查询并映射指定数据点的实时数据。
-     *
-     * @param pointKey   指定的数据点。
-     * @param mapperType 映射器的类型。
-     * @param mapperArgs 映射器的参数。
-     * @return 被映射的指定数据点的实时数据。
-     * @throws ServiceException 服务异常。
-     */
-    List<TimedValue> mappingRealtime(
-            LongIdKey pointKey,
-            String mapperType, Object[] mapperArgs) throws ServiceException;
-
-    /**
-     * 查询指定数据点的持久化数据。
-     *
-     * @param pointKey  指定的数据点。
-     * @param startDate 持久数据的开始时间。
-     * @param endDate   持久数据的结束时间。
-     * @return 指定数据点的持久数据。
-     * @throws ServiceException 服务异常。
-     */
-    List<TimedValue> lookupPersistence(LongIdKey pointKey, Date startDate, Date endDate) throws ServiceException;
-
-    /**
      * 查询并映射指定数据点的持久化数据。
      *
+     * @param mapperType 映射器的类型。
      * @param pointKey   指定的数据点。
      * @param startDate  持久数据的开始时间。
      * @param endDate    持久数据的结束时间。
-     * @param mapperType 映射器的类型。
      * @param mapperArgs 映射器的参数。
      * @return 被映射的指定数据的持久化数据。
      * @throws ServiceException 服务异常。
      */
-    List<TimedValue> mappingPersistence(
-            LongIdKey pointKey, Date startDate, Date endDate,
-            String mapperType, Object[] mapperArgs) throws ServiceException;
+    List<TimedValue> mappingPersistenceValue(
+            String mapperType, LongIdKey pointKey, Date startDate, Date endDate, Object[] mapperArgs)
+            throws ServiceException;
+
+    /**
+     * 查询并映射指定数据点的持久化数据。
+     *
+     * @param mapperType 映射器的类型。
+     * @param pointKey   指定的数据点。
+     * @param startDate  持久数据的开始时间。
+     * @param endDate    持久数据的结束时间。
+     * @param mapperArgs 映射器的参数。
+     * @return 被映射的指定数据的持久化数据。
+     * @throws ServiceException 服务异常。
+     */
+    List<TimedValue> mappingFilteredValue(
+            String mapperType, LongIdKey pointKey, Date startDate, Date endDate, Object[] mapperArgs)
+            throws ServiceException;
+
+    /**
+     * 查询并映射指定数据点的持久化数据。
+     *
+     * @param mapperType 映射器的类型。
+     * @param pointKey   指定的数据点。
+     * @param startDate  持久数据的开始时间。
+     * @param endDate    持久数据的结束时间。
+     * @param mapperArgs 映射器的参数。
+     * @return 被映射的指定数据的持久化数据。
+     * @throws ServiceException 服务异常。
+     */
+    List<TimedValue> mappingTriggeredValue(
+            String mapperType, LongIdKey pointKey, Date startDate, Date endDate, Object[] mapperArgs)
+            throws ServiceException;
 }