+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - com.dwarfeng.fdr.stack.service.FilteredValueMaintainService.previous
   - com.dwarfeng.fdr.stack.service.PersistenceValueMaintainService.previous
   - com.dwarfeng.fdr.stack.service.TriggeredValueMaintainService.previous
+- com.dwarfeng.fdr.stack.service.MappingLookupService 完全重写，功能增强。
 
 #### Bug修复